WIP: Waypoint Icons

+++ npserver/models.py
@@ -86,7 +86,7 @@ class FlightPlan(db.Model):
     owner = relationship("User", back_populates="flightplans")
     map_center_longitude = Column(Float)
     map_center_latitude = Column(Float)
-    map_zoom_level = Column(Integer)
+    map_zoom_level = Column(Float)
     steerpoints = relationship("Coordinate",
                                back_populates="steerpoint_flightplan",
                                foreign_keys="[Coordinate.fp_steerpoint_id]")
